♻️ Refactor: BigDecimal 변수 타입 double로 변경

+++ backend/sleepony/src/main/java/com/c208/sleephony/domain/user/entity/User.java
@@ -2,7 +2,6 @@ package com.c208.sleephony.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -30,10 +29,10 @@ public class User {
     private String nickname;
 
     @Column(name = "height", precision = 4, scale = 1)
-    private BigDecimal height;
+    private double height;
 
     @Column(name = "weight", precision = 4, scale = 1)
-    private BigDecimal weight;
+    private double weight;
 
     @Column(name = "birth_date")
     private LocalDate birthDate;
